refactor: pg 시스템 연동 패키지명 gateway로 변경
