+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/gateway/PgPaymentDto.java
@@ -1,4 +1,4 @@
-package com.loopers.infrastructure.payment;
+package com.loopers.infrastructure.gateway;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
