+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/gateway/PgPaymentFeignClient.java
@@ -1,4 +1,4 @@
-package com.loopers.infrastructure.payment;
+package com.loopers.infrastructure.gateway;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
